opensarm/pyproject.toml
[project]
name = "sarm"
version = "0.1.0"
description = "Reward learning with LeRobot and PyTorch 2.7 + CUDA 12.8"
readme = "README.md"
requires-python = ">=3.10,<3.11"
license = { file = "LICENSE" }

dependencies = [
  "torch==2.9.0",
  "torchvision",
  "torchaudio",
  "torchcodec==0.4",
  "hydra-core==1.2.0",
  "faker>=24.0.0",
  "wandb==0.13.3",
  "omegaconf==2.3.0",
  "datasets==3.6.0",
  "opencv-python==4.7.0.72",
  "moviepy==1.0.3",
  "zarr==2.18.3",
  "decord==0.6.0",


  "tqdm",
  "jsonlines",
  "draccus",
  "av",
  "numpy>=1.22.4,<2.0.0",
  "einops",
  "sentencepiece",
  "transformers",
  "ml_collections",
  "tyro",
  "datasets>=2.0.0",
  "typing-extensions",
  "filelock",
]

[dependency-groups]
dev = [
  "pytest",
  "ruff",
  "pre-commit",
  "ipykernel",
  "matplotlib",
]

[tool.uv]
override-dependencies = []

[tool.uv.sources]
lerobot = { git = "https://github.com/huggingface/lerobot", rev = "0cf864870cf29f4738d3ade893e6fd13fbd7cdb5" }

[tool.uv.workspace]
members = ["packages/*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
  "lerobot",
  "models",
  "workspace",
  "config",
  "train.py",
  "train_utils.py",
  "data_utils.py",
]
